pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "githubkit>=0.12.15",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-client>=0.22.1",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]
